refactor: split execution.rs god-file into modules
