+++ homeboy.json
@@ -642,6 +642,11 @@
         "core_boundary_leak:core-agnostic-source::src/core/runner/capabilities.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `composer` appears at line <line> in behavioral context `top-level`::CoreBoundaryLeak",
         "core_boundary_leak:core-agnostic-source::src/core/runner/capabilities.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `npm` appears at line <line> in behavioral context `top-level`::CoreBoundaryLeak",
         "core_boundary_leak:core-agnostic-source::src/core/runner/capabilities.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `php` appears at line <line> in behavioral context `top-level`::CoreBoundaryLeak",
+        "core_boundary_leak:core-agnostic-source::src/core/runner/execution/tests/policy.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `cargo` appears at line <line> in behavioral context `test_runner_policy_enforces_projects_commands_workspace_and_artifacts`::CoreBoundaryLeak",
+        "core_boundary_leak:core-agnostic-source::src/core/runner/execution/tests/policy.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `rust` appears at line <line> in behavioral context `test_required_extensions_for_command_reads_extension_flags`::CoreBoundaryLeak",
+        "core_boundary_leak:core-agnostic-source::src/core/runner/execution/tests/policy.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `wordpress` appears at line <line> in behavioral context `test_required_extensions_for_command_reads_extension_flags`::CoreBoundaryLeak",
+        "core_boundary_leak:core-agnostic-source::src/core/runner/execution/tests/redaction.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `wordpress` appears at line <line> in behavioral context `runner_exec_failure_error_promotes_homeboy_job_event_message_error`::CoreBoundaryLeak",
+        "core_boundary_leak:core-agnostic-source::src/core/runner/execution/tests/redaction.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `wordpress` appears at line <line> in behavioral context `runner_exec_failure_error_promotes_homeboy_stdout_error`::CoreBoundaryLeak",
         "core_boundary_leak:core-agnostic-source::src/core/runner/lab_capabilities.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `composer` appears at line <line> in behavioral context `lab_runner_capability_contract`::CoreBoundaryLeak",
         "core_boundary_leak:core-agnostic-source::src/core/runner/lab_capabilities.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `npm` appears at line <line> in behavioral context `lab_runner_capability_contract`::CoreBoundaryLeak",
         "core_boundary_leak:core-agnostic-source::src/core/runner/lab_capabilities.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `php` appears at line <line> in behavioral context `lab_runner_capability_contract`::CoreBoundaryLeak",
@@ -692,7 +697,8 @@
         "field_patterns::src/core/git/pr_policy.rs::RepeatedFieldPattern",
         "field_patterns::src/core/observation/records.rs::RepeatedFieldPattern",
         "field_patterns::src/core/observation/runs_service.rs::RepeatedFieldPattern",
-        "field_patterns::src/core/runner/execution.rs::RepeatedFieldPattern",
+        "field_patterns::src/core/runner/execution/daemon.rs::RepeatedFieldPattern",
+        "field_patterns::src/core/runner/execution/mod.rs::RepeatedFieldPattern",
         "field_patterns::src/core/runner/mod.rs::RepeatedFieldPattern",
         "field_patterns::src/core/runner/session.rs::RepeatedFieldPattern",
         "field_patterns::src/core/server/mod.rs::RepeatedFieldPattern",
@@ -723,8 +729,9 @@
         "intra-method-duplication::src/core/quality.rs::IntraMethodDuplicate",
         "intra-method-duplication::src/core/rig/pipeline.rs::IntraMethodDuplicate",
         "intra-method-duplication::src/core/runner/connection.rs::IntraMethodDuplicate",
-        "intra-method-duplication::src/core/runner/execution.rs::IntraMethodDuplicate",
+        "intra-method-duplication::src/core/runner/execution/daemon_api.rs::IntraMethodDuplicate",
         "intra-method-duplication::src/core/runner/execution/extension_parity.rs::IntraMethodDuplicate",
+        "intra-method-duplication::src/core/runner/execution/failure.rs::IntraMethodDuplicate",
         "intra-method-duplication::src/core/runner/git_dependency_materialization.rs::IntraMethodDuplicate",
         "intra-method-duplication::src/core/runner/validation_dependencies.rs::IntraMethodDuplicate",
         "intra-method-duplication::src/core/runner/worker.rs::IntraMethodDuplicate",
@@ -799,8 +806,6 @@
         "structural::src/core/rig/pipeline.rs::HighItemCount",
         "structural::src/core/runner/connection.rs::HighItemCount",
         "structural::src/core/runner/evidence.rs::HighItemCount",
-        "structural::src/core/runner/execution.rs::GodFile",
-        "structural::src/core/runner/execution.rs::HighItemCount",
         "structural::src/core/runner/lab/agent_task_bridge.rs::HighItemCount",
         "structural::src/core/runner/lab/offload.rs::GodFile",
         "structural::src/core/runner/lab_args/tests.rs::HighItemCount",
@@ -822,6 +827,9 @@
         "test_quality::src/commands/runner/doctor/tests.rs::VacuousTest",
         "test_quality::src/core/agent_task_scheduler/tests/outcome_tests.rs::VacuousTest",
         "test_quality::src/core/code_audit/duplication/tests.rs::VacuousTest",
+        "test_quality::src/core/runner/execution/tests/policy.rs::VacuousTest",
+        "test_quality::src/core/runner/execution/tests/redaction.rs::VacuousTest",
+        "test_quality::src/core/runner/execution/tests/secret_source.rs::VacuousTest",
         "test_quality::tests/command_surface_test.rs::VacuousTest",
         "thin_command_adapter::src/commands/agent_task/tool.rs::ThinCommandAdapterViolation",
         "thin_command_adapter::src/commands/bench/observation.rs::ThinCommandAdapterViolation",
